Remove unused imports and move files
