+++ apps/docs/app/root/layout/components/index.ts
@@ -1,0 +1,2 @@
+export { default as SearchResults } from "./SearchResults";
+export { default as GlobalSearchInput } from "./GlobalSearchInput";
