+++ apps/docs/app/routes/_base/content-menu/SearchableContentMenu.tsx
@@ -1,11 +1,6 @@
-import { useLocation, useNavigate } from "@remix-run/react";
 import { Box } from "@vygruppen/spor-react";
-import { matchSorter } from "match-sorter";
 import React from "react";
-import { useMenu } from "~/utils/useMenu";
 import { ContentMenu } from "./ContentMenu";
-import { GlobalSearchInput } from "./GlobalSearchInput";
-import { SearchResults } from "./SearchResults";
 
 export const SearchableContentMenu = () => {
   const focusableRef = React.useRef<HTMLButtonElement>(null);
